+++ opentelemetry-api/CHANGELOG.md
@@ -8,6 +8,8 @@
   ([#1118](https://github.com/open-telemetry/opentelemetry-python/pull/1118))
 - Allow for Custom Trace and Span IDs Generation - `IdsGenerator` for TracerProvider
   ([#1153](https://github.com/open-telemetry/opentelemetry-python/pull/1153))
+- Parent is now always passed in via Context, intead of Span or SpanContext
+  ([#1146](https://github.com/open-telemetry/opentelemetry-python/pull/1146))
 
 ## Version 0.13b0